split up cpu topology class
